# opcache: optimize some more functions on compile time