--- a/ext/opcache/Optimizer/sccp.c
+++ b/ext/opcache/Optimizer/sccp.c
@@ -844,19 +844,16 @@ static inline int ct_eval_func_call(
 				|| zend_string_equals_literal(name, "urldecode")
 				|| zend_string_equals_literal(name, "rawurlencode")
 				|| zend_string_equals_literal(name, "rawurldecode")
-				|| zend_string_equals_literal(name, "php_uname")
+				|| zend_string_equals_literal(name, "strtoupper")
+				|| zend_string_equals_literal(name, "strtolower")
 				|| zend_string_equals_literal(name, "dirname")
-				|| zend_string_equals_literal(name, "basename")
-				|| zend_string_equals_literal(name, "md5")
-				|| zend_string_equals_literal(name, "crc32")
-				|| zend_string_equals_literal(name, "sha1")) {
+				|| zend_string_equals_literal(name, "crc32")) {
 			if (Z_TYPE_P(args[0]) != IS_STRING) {
 				return FAILURE;
 			}
 			/* pass */
 		} else if (zend_string_equals_literal(name, "array_keys")
 				|| zend_string_equals_literal(name, "array_values")
-				|| zend_string_equals_literal(name, "array_unique")
 				|| zend_string_equals_literal(name, "array_filter")) {
 			if (Z_TYPE_P(args[0]) != IS_ARRAY) {
 				return FAILURE;
@@ -874,7 +871,8 @@ static inline int ct_eval_func_call(
 				}
 			} ZEND_HASH_FOREACH_END();
 			/* pass */
-		} else if (zend_string_equals_literal(name, "implode")) {
+		} else if (zend_string_equals_literal(name, "implode")
+                || zend_string_equals_literal(name, "array_unique")) {
 			zval *entry;
 
 			if (Z_TYPE_P(args[0]) != IS_ARRAY) {
@@ -975,8 +973,7 @@ static inline int ct_eval_func_call(
 				|| zend_string_equals_literal(name, "str_contains")
 				|| zend_string_equals_literal(name, "str_starts_with")
 				|| zend_string_equals_literal(name, "str_ends_with")
-				|| zend_string_equals_literal(name, "version_compare")
-				|| zend_string_equals_literal(name, "basename")) {
+				|| zend_string_equals_literal(name, "version_compare")) {
 			if (Z_TYPE_P(args[0]) != IS_STRING
 					|| Z_TYPE_P(args[1]) != IS_STRING) {
 				return FAILURE;
@@ -1018,8 +1015,7 @@ static inline int ct_eval_func_call(
 				}
 			}
 			/* pass */
-		} else if (zend_string_equals_literal(name, "version_compare")
-				|| zend_string_equals_literal(name, "str_replace")) {
+		} else if (zend_string_equals_literal(name, "str_replace")) {
 			if (Z_TYPE_P(args[0]) != IS_STRING
 					|| Z_TYPE_P(args[1]) != IS_STRING
 					|| Z_TYPE_P(args[2]) != IS_STRING) {