--- a/ext/opcache/Optimizer/sccp.c
+++ b/ext/opcache/Optimizer/sccp.c
@@ -844,8 +844,7 @@ static inline int ct_eval_func_call(
 				|| zend_string_equals_literal(name, "urldecode")
 				|| zend_string_equals_literal(name, "rawurlencode")
 				|| zend_string_equals_literal(name, "rawurldecode")
-				|| zend_string_equals_literal(name, "strtoupper")
-				|| zend_string_equals_literal(name, "strtolower")
+				|| zend_string_equals_literal(name, "php_uname")
 				|| zend_string_equals_literal(name, "dirname")
 				|| zend_string_equals_literal(name, "crc32")) {
 			if (Z_TYPE_P(args[0]) != IS_STRING) {
@@ -872,7 +871,7 @@ static inline int ct_eval_func_call(
 			} ZEND_HASH_FOREACH_END();
 			/* pass */
 		} else if (zend_string_equals_literal(name, "implode")
-                || zend_string_equals_literal(name, "array_unique")) {
+				|| zend_string_equals_literal(name, "array_unique")) {
 			zval *entry;
 
 			if (Z_TYPE_P(args[0]) != IS_ARRAY) {
@@ -916,6 +915,11 @@ static inline int ct_eval_func_call(
 					|| (Z_TYPE_P(args[1]) != IS_LONG)) {
 				return FAILURE;
 			}
+			if (Z_LVAL_P(args[1]) < 1) {
+				// levels must be >= 1, else we get a ValueError
+				return FAILURE;
+			}
+
 			/* pass */
 		} else if (zend_string_equals_literal(name, "trim")
 				|| zend_string_equals_literal(name, "rtrim")
@@ -1015,6 +1019,7 @@ static inline int ct_eval_func_call(
 				}
 			}
 			/* pass */
+		// todo: version_compare with 3 arguments got removed, as we should add a proper check for the comperator, else we hide a ValueError
 		} else if (zend_string_equals_literal(name, "str_replace")) {
 			if (Z_TYPE_P(args[0]) != IS_STRING
 					|| Z_TYPE_P(args[1]) != IS_STRING