I profiled our production code and found some (more or less rare) function calls which could get evaluated on compile time already + extended it by some other function calls which should always return a static result.

**List: (updated one)**
- dirname(string[, int]) 
- array_unique(array)
- array_filter(array)
- crc32(string)
- str_replace(string, string, string)


Dummy code:
```
// source from official docs: https://www.php.net/manual/en/function.php-uname.php
if (strtoupper(substr(PHP_OS, 0, 3)) === 'WIN') {
    echo 'This is a server using Windows!';
} else {
    echo 'This is a server not using Windows!';
}

function getRoot()
{
    return dirname(__DIR__);
}

function getBaseCacheKey()
{
    return str_replace('a', 'z', md5(PHP_VERSION));
}

class A {
    public const BASE_ABILITIES = [
        1,
        2,
    ];
}
class B extends A {
    public const ABILITIES = [
        2,
        3,
        4,
        0,
    ];

    public function getAllAbilities(): string
    {
        return implode(',', array_filter(
            array_unique(array_merge(self::BASE_ABILITIES, self::ABILITIES)))
        );
    }
}
```

**Opcodes before:**
```
$_main:
     ; (lines=9, args=0, vars=0, tmps=2)
     ; (after optimizer)
     ; /www/php-src/test.php:1-144
0000 INIT_FCALL 1 96 string("strtoupper")
0001 SEND_VAL string("Lin") 1
0002 V1 = DO_ICALL
0003 T0 = IS_IDENTICAL V1 string("WIN")
0004 JMPZ T0 0007
0005 ECHO string("This is a server using Windows!")
0006 RETURN int(1)
0007 ECHO string("This is a server not using Windows!")
0008 RETURN int(1)

getRoot:
     ; (lines=1, args=0, vars=0, tmps=0)
     ; (after optimizer)
     ; /www/php-src/test.php:8-11
0000 RETURN string("/www")

getBaseCacheKey:
     ; (lines=10, args=0, vars=0, tmps=1)
     ; (after optimizer)
     ; /www/php-src/test.php:13-16
0000 INIT_FCALL 3 128 string("str_replace")
0001 SEND_VAL string("a") 1
0002 SEND_VAL string("z") 2
0003 INIT_FCALL 1 96 string("md5")
0004 SEND_VAL string("8.0.0-dev") 1
0005 V0 = DO_ICALL
0006 SEND_VAR V0 3
0007 V0 = DO_ICALL
0008 VERIFY_RETURN_TYPE V0
0009 RETURN V0
LIVE RANGES:
     0: 0008 - 0009 (tmp/var)

B::getAllAbilities:
     ; (lines=11, args=0, vars=0, tmps=1)
     ; (after optimizer)
     ; /www/php-src/test.php:32-37
0000 INIT_FCALL 2 112 string("implode")
0001 SEND_VAL string(",") 1
0002 INIT_FCALL 1 96 string("array_filter")
0003 INIT_FCALL 1 96 string("array_unique")
0004 SEND_VAL array(...) 1
0005 V0 = DO_ICALL
0006 SEND_VAR V0 1
0007 V0 = DO_ICALL
0008 SEND_VAR V0 2
0009 V0 = DO_ICALL
0010 RETURN V0
```

**Opcodes afterwards:**
```
$_main:
     ; (lines=2, args=0, vars=0, tmps=0)
     ; (after optimizer)
     ; /www/php-src/test.php:1-159
0000 ECHO string("This is a server not using Windows!")
0001 RETURN int(1)


getRoot:
     ; (lines=1, args=0, vars=0, tmps=0)
     ; (after optimizer)
     ; /www/php-src/test.php:8-11
0000 RETURN string("/www")

getBaseCacheKey:
     ; (lines=1, args=0, vars=0, tmps=0)
     ; (after optimizer)
     ; /www/php-src/test.php:13-16
0000 RETURN string("0583081z712641ez72e8bb3fz6d3d645")

B::getAllAbilities:
     ; (lines=2, args=0, vars=0, tmps=1)
     ; (after optimizer)
     ; /www/php-src/test.php:32-37
0000 V0 = QM_ASSIGN string("1,2,3,4")
0001 RETURN V0
```