@GrahamCampbell 2020-07-18
I am not sure those hash functions are a good idea to optimise. sha1 and md5 have no use as crypto hashes and are much slower than modern hashing algorithms if you just want a quick hash with good distribution properties.