@GrahamCampbell 2020-07-18
I am :+1: for `array_filter`, but perhaps not any of the others, actually. array_unique is already a slow function (high complexity), and so you gain nothing in reality from optimising it in this way unless your inputs are of size 0 or 1.