@GrahamCampbell 2020-07-18
If you have constant stuff being passed into array_unique, I'd question why that code is like that, more than seeing if we can optimize it.