@TysonAndre 2020-07-19
> I am :+1: for `array_filter`, but perhaps not any of the others, actually. `array_unique` is already a slow function (high complexity), and so you gain nothing in reality from optimising it in this way unless your inputs are of size 0 or 1. If you have constant stuff being passed into array_unique, I'd question why that code is like that, more than seeing if we can optimize it. Similarly for the others, having constant input seems like not a practical or common use case?



In PHP 8 and probably 7, I think that array_unique with 2 args is of complexity `O(n)`, not of complexity `O(n*n)`.
The default sort type is SORT_STRING. It inserts them all into a HashTable by the numeric or string key and only does that one O(1) operation per element.

See ext/standard/array for `PHP_FUNCTION(array_unique)`

I'd agree that code needing to use array_unique is uncommon, but not impossible. For example, `self::$privateValidTypes = array_values(array_unique(array_merge(self::TYPES_A, self::TYPES_B)));`

(timing strangeness possibly due to cache locality or the time being amortized and not accounting for resizing, but it isn't quadratic)

```
php > function benchmark_unique(int $n) { $x = range(1, $n); $start = microtime(true); $y = array_unique($x); $end = microtime(true); printf("Elapsed: %.3f\n", $end - $start);}
php > benchmark_unique(100000);
Elapsed: 0.014
php > benchmark_unique(200000);
Elapsed: 0.023
php > benchmark_unique(300000);
Elapsed: 0.036
php > benchmark_unique(500000);
Elapsed: 0.051
```