@brainexe 2020-07-23
Onother live example for the dirname() optimization: [Symfony dumped DIC](https://github.com/symfony/symfony/blob/cedb5cd/src/Symfony/Component/DependencyInjection/Tests/Fixtures/php/services_inline_requires.php#L86)
-> in our project we have a few dozen such service calls which are all load `dirname(__DIR__, 1)`