@GrahamCampbell 2020-07-23
I think your updated proposed list of functions looks much better, now. 👍