@nikic 2020-11-27
Sorry for the delay! I finally got around to rewriting this code, as the current approach seemed unnecessarily fragile (see the version_compare case found here). I have added support for throwing functions (we will simply discard the exception), so we only need to check for functions/args that are runtime dependent or can result in warnings, now done in this function: https://github.com/php/php-src/blob/99a3dbce819dcc0d3e13a49a1aa9a5699296298b/ext/opcache/Optimizer/sccp.c#L784

Would you mind rebasing on top of these changes? I think it should now be just a matter of adding functions to the list, without the need to add any new logic.