@TysonAndre 2020-11-27
> Would you mind rebasing on top of these changes? I think it should now be just a matter of adding functions to the list, without the need to add any new logic.

Those changes introduced some differences in behavior due to opcache always evaluating the code with `strict_types=0` - https://github.com/php/php-src/commit/0ce9b5f7550ac89c59e541549641512168115594#r44598985 , but look good if that was fixed

------

Obviously, logic for array_filter would need to be kept to check if there were 1 arguments - the callback has side effects

The argument counts/types for str_replace should also continue to need to be checked for arrays within arrays because those would emit notices during compilation

```
php > var_export(str_replace(['foo'], [[]], 'foo'));

Notice: Array to string conversion in php shell code on line 1
'Array'
```