@nikic 2021-07-19
I've added array_unique and str_replace from here in https://github.com/php/php-src/commit/273720dcf25069504e5f5eb81242de9498bd88bd. dirname has already been added earlier (for non-Windows).